requirements.txt
Django==1.4.3
PIL==1.1.7
distribute==0.6.34
dj-database-url==0.2.1
django-haystack==1.2.7
django-registration==0.8
gunicorn==0.17.2
psycopg2==2.4.6
pysolr==3.0.5
requests==1.2.0
wsgiref==0.1.2
